[red] project file path 변경
