+++ SSDProject/SSDTest/SSDTest.vcxproj
@@ -106,6 +106,7 @@
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">Create</PrecompiledHeader>
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Release|x64'">Create</PrecompiledHeader>
     </ClCompile>
+    <ClCompile Include="TestShellApplication.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\SSDProject\SSDProject.vcxproj">
